feat(home): separated form and list presentational components
