+++ src/components/TodoList/index.js
@@ -1,0 +1,1 @@
+export { default as TodoList } from "./TodoList";
